src/media_system/featured_image.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Created on Tue Nov 26 01:41:57 2024

@author: thesaint
"""

# src/media_system/featured_image.py
from typing import Dict, Optional
import logging
from pathlib import Path
from datetime import datetime
from PIL import Image
from io import BytesIO

class FeaturedImageHandler:
    def __init__(self, image_optimizer):
        self.image_optimizer = image_optimizer
        self.featured_size = (1200, 630)  # Optimal size for social sharing
        self.setup_logging()
        
    def setup_logging(self):
        log_dir = Path('logs')
        log_dir.mkdir(exist_ok=True)
        logging.basicConfig(
            filename=log_dir / f'featured_image_{datetime.now():%Y%m%d}.log',
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s - %(message)s'
        )
        
    async def process_featured_image(self, image_data: Dict) -> Optional[Dict]:
        try:
            # Optimize image for featured use
            optimized = await self.image_optimizer.optimize_image(
                image_data['data'],
                image_data['id']
            )
            
            # Additional featured image processing
            featured = await self._prepare_featured_image(optimized)
            
            return {
                'id': image_data['id'],
                'data': featured['data'],
                'size': self.featured_size,
                'alt_text': self._generate_alt_text(image_data),
                'caption': self._generate_caption(image_data)
            }
        except Exception as e:
            logging.error(f"Featured image processing failed: {str(e)}")
            return None